Build new proposals and commit flows
