--- src/client/actions/index.ts
@@ -1,2 +1,0 @@
-export * from "./propose-add-member.js";
-export * from "./propose-update-metadata.js";
