+++ src/client/group/index.ts
@@ -1,0 +1,2 @@
+export * as Proposals from "./proposals/index.js";
+export * from "./marmot-group.js";
