+++ src/client/group/proposals/index.ts
@@ -1,0 +1,2 @@
+export * from "./add-member.js";
+export * from "./update-metadata.js";
