+++ src/client/group/proposals/update-metadata.ts
@@ -1,0 +1,32 @@
+import { type ProposalGroupContextExtensions } from "ts-mls/proposal.js";
+import { replaceExtension } from "../../../core/extensions.js";
+import { marmotGroupDataToExtension } from "../../../core/marmot-group-data.js";
+import { MarmotGroupData } from "../../../core/protocol.js";
+import type { ProposalBuilder } from "../marmot-group.js";
+
+/** Builds a proposal to update a group's marmot group data extension */
+export function proposeUpdateMetadata(
+  metadata: Partial<MarmotGroupData>,
+): ProposalBuilder<ProposalGroupContextExtensions> {
+  return async ({ state, groupData }) => {
+    // Create updated group data, preserving existing values for unchanged fields
+    const updatedGroupData: MarmotGroupData = {
+      ...groupData,
+      ...metadata,
+    };
+
+    // Convert to MLS extension
+    const updatedExtension = marmotGroupDataToExtension(updatedGroupData);
+
+    return {
+      proposalType: "group_context_extensions",
+      groupContextExtensions: {
+        // Replace the marmot group data extension with the updated one
+        extensions: replaceExtension(
+          state.groupContext.extensions,
+          updatedExtension,
+        ),
+      },
+    };
+  };
+}
