+++ src/client/index.ts
@@ -1,3 +1,2 @@
 export * from "./interfaces.js";
 export * from "./marmot-client.js";
-export * as Transactions from "./transactions/index.js";
