+++ src/client/marmot-client.ts
@@ -19,7 +19,7 @@ import { generateKeyPackage } from "../core/key-package.js";
 import { GroupStore } from "../store/group-store.js";
 import { KeyPackageStore } from "../store/key-package-store.js";
 import { NostrPool } from "./interfaces.js";
-import { MarmotGroup } from "./marmot-group.js";
+import { MarmotGroup } from "./group/marmot-group.js";
 
 export type MarmotClientOptions = {
   /** The signer used for the clients identity */
