--- src/client/transactions/index.ts
@@ -1,1 +1,0 @@
-export * from "./accept-add-member.js";
